refactor: extract shared SectionHeader/DetailRow and simplify route files
Extract duplicated SectionHeader and DetailRow components into a shared
detail-layout.tsx UI component used by the agent detail and job detail
routes. Also simplify navigation callbacks and template literals.
